refactor: ExperienceSection component📦
